+++ textprofilerfrontend/src/App.svelte
@@ -122,7 +122,6 @@
   <div class="grow" />
 
   <Search
-    mainDatasetName={datasetInfo?.name}
     columnNames={datasetInfo?.column_info
       .filter((col) => col.type === "text")
       .map((col) => col.name)}
